Use unsigned long for ee_ptr_int.
This allows 32-bit systems to work.

+++ freebsd/core_portme.h
@@ -100,7 +100,7 @@ typedef signed int ee_s32;
 typedef double ee_f32;
 typedef unsigned char ee_u8;
 typedef unsigned int ee_u32;
-typedef unsigned long long ee_ptr_int;
+typedef unsigned long ee_ptr_int;
 typedef size_t ee_size_t;
 /* align an offset to point to a 32b value */
 #define align_mem(x) (void *)(4 + (((ee_ptr_int)(x) - 1) & ~3))
